part2/functions/remediator-agent/requirements.txt
# Same deps as diagnoser; shared agents rely on these
google-cloud-pubsub
google-cloud-secret-manager
requests
openai
anthropic